# cli/generated/ramekin-client/src/models/create_tag_request.rs
/*
 * ramekin-server
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 * Generated by: https://openapi-generator.tech
 */

use crate::models;
use serde::{Deserialize, Serialize};

#[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
pub struct CreateTagRequest {
    #[serde(rename = "name")]
    pub name: String,
}

impl CreateTagRequest {
    pub fn new(name: String) -> CreateTagRequest {
        CreateTagRequest { name }
    }
}