# cli/generated/ramekin-client/src/models/tag_item.rs
/*
 * ramekin-server
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 * Generated by: https://openapi-generator.tech
 */

use crate::models;
use serde::{Deserialize, Serialize};

#[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
pub struct TagItem {
    #[serde(rename = "created_at")]
    pub created_at: String,
    #[serde(rename = "id")]
    pub id: uuid::Uuid,
    #[serde(rename = "name")]
    pub name: String,
}

impl TagItem {
    pub fn new(created_at: String, id: uuid::Uuid, name: String) -> TagItem {
        TagItem {
            created_at,
            id,
            name,
        }
    }
}