# cli/generated/ramekin-client/src/models/tags_list_response.rs
/*
 * ramekin-server
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 * Generated by: https://openapi-generator.tech
 */

use crate::models;
use serde::{Deserialize, Serialize};

#[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
pub struct TagsListResponse {
    #[serde(rename = "tags")]
    pub tags: Vec<models::TagItem>,
}

impl TagsListResponse {
    pub fn new(tags: Vec<models::TagItem>) -> TagsListResponse {
        TagsListResponse { tags }
    }
}