# ramekin-ui/generated-client/models/CreateTagRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ramekin-server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateTagRequest
 */
export interface CreateTagRequest {
    /**
     * 
     * @type {string}
     * @memberof CreateTagRequest
     */
    name: string;
}

/**
 * Check if a given object implements the CreateTagRequest interface.
 */
export function instanceOfCreateTagRequest(value: object): value is CreateTagRequest {
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function CreateTagRequestFromJSON(json: any): CreateTagRequest {
    return CreateTagRequestFromJSONTyped(json, false);
}

export function CreateTagRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateTagRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
    };
}

export function CreateTagRequestToJSON(json: any): CreateTagRequest {
    return CreateTagRequestToJSONTyped(json, false);
}

export function CreateTagRequestToJSONTyped(value?: CreateTagRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
    };
}