# ramekin-ui/generated-client/models/TagsListResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ramekin-server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TagItem } from './TagItem';
import {
    TagItemFromJSON,
    TagItemFromJSONTyped,
    TagItemToJSON,
    TagItemToJSONTyped,
} from './TagItem';

/**
 * 
 * @export
 * @interface TagsListResponse
 */
export interface TagsListResponse {
    /**
     * 
     * @type {Array<TagItem>}
     * @memberof TagsListResponse
     */
    tags: Array<TagItem>;
}

/**
 * Check if a given object implements the TagsListResponse interface.
 */
export function instanceOfTagsListResponse(value: object): value is TagsListResponse {
    if (!('tags' in value) || value['tags'] === undefined) return false;
    return true;
}

export function TagsListResponseFromJSON(json: any): TagsListResponse {
    return TagsListResponseFromJSONTyped(json, false);
}

export function TagsListResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): TagsListResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'tags': ((json['tags'] as Array<any>).map(TagItemFromJSON)),
    };
}

export function TagsListResponseToJSON(json: any): TagsListResponse {
    return TagsListResponseToJSONTyped(json, false);
}

export function TagsListResponseToJSONTyped(value?: TagsListResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'tags': ((value['tags'] as Array<any>).map(TagItemToJSON)),
    };
}